+++ src/utils/api/open-api/shape-swagger.js
@@ -25,4 +25,4 @@ Object.entries(dfn.paths).forEach(([apiPath, verbs]) => {
   })
 })
 
-exports.methods = methods
+module.exports = methods
